# docs/architecture/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: UnFranchise Marketing App API
  description: |
    API for the UnFranchise Marketing App - A corporate-controlled content sharing
    and engagement platform for UnFranchise Owners (UFOs).

    ## Authentication
    Most endpoints require Bearer token authentication via JWT.
    Include the access token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes
    - Standard API endpoints: 100 requests per minute
    - Share endpoints: 500 requests per hour
    - Upload endpoints: 20 requests per hour

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Pagination
    List endpoints support pagination via query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sort`: Sort field with optional `-` prefix for descending (e.g., `-createdAt`)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@unfranchise.com
  license:
    name: Proprietary
servers:
  - url: https://api.unfranchise.com/v1
    description: Production server
  - url: https://staging-api.unfranchise.com/v1
...
paths:
...
  /content/featured:
    get:
      tags: [Content]
      summary: Get featured content
      description: Get list of featured content items
      operationId: getFeaturedContent
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Featured content retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentItem'
                  meta:
                    $ref: '#/components/schemas/Meta'
...
  /shares:
    post:
      tags: [Sharing]
      summary: Create share
      description: Share content via SMS, email, or social
      operationId: createShare
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareCreate'
      responses:
        '201':
          description: Share created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ShareResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Sharing]
      summary: Get share history
      description: Get user's share history with pagination
      operationId: getShareHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: channel
          in: query
          schema:
            type: string
            enum: [sms, email, social]
        - name: contentId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ShareEvent'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  meta:
                    $ref: '#/components/schemas/Meta'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # User & Auth
...
    ContentItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          format: uri
        contentType:
          type: string
          enum: [video, image, pdf, landing_page, product, bundle, share_card]
        category:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
        tags:
          type: array
          items:
            type: string
        channels:
          type: array
          items:
            type: string
            enum: [sms, email, social]
        market:
          type: string
        language:
          type: string
        publishStatus:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ShareCreate:
      type: object
      required: [contentId, channel, recipients]
      properties:
        contentId:
          type: string
          format: uuid
        channel:
          type: string
          enum: [sms, email, social]
        recipients:
          type: array
          items:
            type: object
            properties:
              contactId:
                type: string
                format: uuid
              email:
                type: string
                format: email
              phone:
                type: string
              name:
                type: string
        personalNote:
          type: string
          maxLength: 500
        sendMethod:
          type: string
          enum: [system, manual]
          default: manual
    ShareResponse:
      type: object
      properties:
        shareId:
          type: string
          format: uuid
        trackingLinks:
          type: array
          items:
            type: object
            properties:
              recipientId:
                type: string
                format: uuid
              shortUrl:
                type: string
                format: uri
              fullUrl:
                type: string
                format: uri
        messagePreview:
          type: string
        status:
          type: string
    ShareEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contentTitle:
          type: string
        contentThumbnail:
          type: string
          format: uri
        channel:
          type: string
        recipientCount:
          type: integer
        clicks:
          type: integer
        views:
          type: integer
        createdAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...